+++ tiledesk-client/index.js
@@ -654,6 +654,7 @@ class TiledeskClient {
       TiledeskClient.myrequest(
         HTTPREQUEST,
         (err, resbody) => {
+          
           if (err) {
             if (callback) {
               callback(err);
@@ -2395,7 +2396,6 @@ class TiledeskClient {
       }
     })
     .catch(function (error) {
-      console.error("An error occurred:", error);
       if (callback) {
         callback(error, null, null);
       }